fix: instabilities with using getResource and gradle caching
